backend/requirements.txt
python-socketio
fastapi
uvicorn
python-dotenv
requests # implicite pour le client io (c'était pas clair)
core-archi-websocket-kube-conceplog